+++ docs/src/conf.py
@@ -20,7 +20,7 @@ copyright = "2024, SKAO"
 author = "Team Wombat"
 
 # The full version, including alpha/beta/rc tags
-release = "0.1.0"
+release = "0.1.1"
 
 
 # -- General configuration ---------------------------------------------------