+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Unreleased
     - WOM-446, WOM-464: Added missing commands to the ``Command`` enum.
     - WOM-464: Added ``ServerStatus.CurrentTime`` key to ``nodes`` and ``attributes`` properties that reads the server's local time.
     - WOM-506: Updated the command authority checks to not block sending a 'TakeAuth' or 'ReleaseAuth' command under any circumstance.
+    - WOM-454 Integrate ska-mid-wms, adding sensors to attributes dict and enable weather station subscriptions.
 - Documentation:
     - KAR-1198: Updated 'How to use SCU'.
 - Bug fixes: